v0.7 Доделал счетчик до сессии ToDo: красиво оформить
