+++ app/src/main/res/values/strings.xml
@@ -30,5 +30,6 @@
     <string name="Lab">Лабораторная</string>
     <string name="title_activity_notes_activity">Notes_activity</string>
     <string name="title_activity_rasp">Rasp</string>
+    <string name="examTime">Счетчик до сессии</string>
 
 </resources>
